Ignore files produced in the staging directory (#12)
* Ignore files produced in the staging directory

Not happy with this yet.  I want the staging dir and its README.md to be part of the github-controlled files, but not any files we may produce in there.  Is that possible? I looked for regex that would only allow README.md but no luck.  So I added a few likely candidates of filetypes.  Please improve.

* Update .gitignore

Fix the regex