Create both pixel and gridline version during downsamling
While the master file cannot be changed from one registration to another without severe loss of the highest frequencies, for the downsampling we can create both pixel and gridline versions of whatever the master file was.  This is helpful when you need to coregistered other data sets, especially images which usually are pixel registered - in contrast, lots of geophysical data are gridline registered.