Update data server file and link script (#43)
We decide that for GMT 6.1 any earth_relief file obtained by a link with the old name should be written to teh server/earth_relief directory on the local machine.  This is now reflected in the table.  The srv_setbackwards_links.sh removes any old links and creates them.  This script is only meant to be run manually, probably just once, but its precense is a log of what happened.