Add two example custom symbols
As part as an enhancement to GMT, I am adding two simple custom symbols that indicates the direction of a Cartesian angle (CCW from the horizontal) or a geographic azimuth (CW from north), with the reference line, angle arc, and annotation added.  These two symbols will be used in a new figure in the custom symbol section.