Really, finalize the srtm coverage script
Learned that it must be gridline registered due to GMT code so we must stay backwards compatible.  DOcumented this better both in MAINTENANCE.md as well as the reasoning in the src_srtm_coverage.sh script.  Also decided that it made more sense to use 0 = no tile, 1 = partial tile, 2 = full tile.