Optimize planetry recipes for keeping highest resolution
See #171 for background.  I have solved the issue by determining the exact increment and tile size required for the highest resolution, which gets a name based on the floor of the actual increment.  Need floor since 2.56 cannot be rint to 3 since we also will make a 03m grid.
Testing this now but seems to work so want to get these into the repo for further testing.