Make sure rounded resolution has 2 digits
When the original data set has an odd resolution like 52,074555234s we floor that to get 52.  However, that code would turn 2.45666 into 2 but we need 02.

This PR set s the required format so that we get a 2-digit resolution value.  With that it looks like the scripts and recipes works for the new candidates Moon, Mercury, Mars, Venus, and Pluto.