Update server-info in Makefile (#179)
See details in #178. The idea is to update the server-info section of Makefile.

I reeplace earth_*_server.txt for *_*_server.txt. The problem with this is that reads the gmt_data_server.txt file. 
To avoid this, first I remove the file. Then, I save the file in the temporal folder (/tmp/) .
Finally, I move the file to the information folder.

With these changes, I think that will get a gmt_data_server.txt for any type cellestial body (earth, moon, mars, etc).