Increase filter width a bit for smallest increments
This PR scales the amount we add to the filter width raw calculation by a factor that is a bit larger for tiny resolutions and drops toward zero for the coarser ones.  Using 250 m as the offset scaled by

FW_ADD=$(gmt math -Q ${FW_OFFSET} ${INC} DIV 0.0166666666667 MUL 0.001 MUL =)

This removed the NaNs in 01m Mercury and Pluto.  The PR also checks if the filtered output has any NaNs, which will alert us to further tinkering with the filter width.