Handle the odd comments with data for pre-6.5 GMT
Due to the 16 bytes needed in gmt_data_server.txt for the Pluto and Mercury resolutions they are flagged as comments starting with #% . That we can handle in 6.5 but pre-6.5 does not know about these so the count of all data sets is off by 2.  This PR subtracts the number of such data sets found from the total datasets given in the first record.  GMT 6.5 can then add back the number of high-res data found.