Many changes in order to build and maintain gmtserver's 4 servers of data (#232)
* Increase filter width a bit for smallest increments

This PR scales the amount we add to the filter width raw calculation by a factor that is a bit larger for tiny resolutions and drops toward zero for the coarser ones.  Using 250 m as the offset scaled by

FW_ADD=$(gmt math -Q ${FW_OFFSET} ${INC} DIV 0.0166666666667 MUL 0.001 MUL =)

This removed the NaNs in 01m Mercury and Pluto.  The PR also checks if the filtered output has any NaNs, which will alert us to further tinkering with the filter width.

* Update srv_downsampler_grid.sh

* More testing

* polish

* Add similar checks and FW to images

* Update srv_downsampler_grid.sh

Settling on final filter scheme - will be added to new DOWNSAMPLING.md soon

* finalize filtering setup

* Handle the odd comments with data for pre-6.5 GMT

Due to the 16 bytes needed in gmt_data_server.txt for the Pluto and Mercury resolutions they are flagged as comments starting with #% . That we can handle in 6.5 but pre-6.5 does not know about these so the count of all data sets is off by 2.  This PR subtracts the number of such data sets found from the total datasets given in the first record.  GMT 6.5 can then add back the number of high-res data found.

* Update srv_do_updates.sh

* Fix tiler script

* Update filter_width_from_output_spacing.sh

* Make -x a script option for downsampling