WIP Fixing missing nodes in mss
Here is a first cut.  Suggest you use this recipe, add a temporary exit in srv_downsamples_grid.c around line 142 and run it as

bash -xv scripts/srv_downsamples_grid.c  earth_mss

amd examine output.  The idea here is that (unfortunately) they called it *.nc so we are stuck with what SRC_CUSTOM and SRC_PROCESS allow. We want

make an *orig* copy if none exists in staging (this will happen once)
RUn the custom command that reads the orig file and expands to -Rd with NaNs

The rest of the scripts should then handle things.