Script to count files on a server or your staging area
Run scripts/srv_staging_vs_candidate.sh staging to get a total cound of the number of files underthat dir, and run the same script with the path to the relevant server dir on the gmtserver.  if the counts differ then probably the scp job timed up due to excessive Internet traffic!